refactor: userをentities layerに移動
