+++ src/_pages/lobby-detail/model/use-chat.tsx
@@ -2,7 +2,7 @@
 
 import { type FormEvent, useEffect, useRef, useState } from "react";
 import { useLobby } from "@/_pages/lobby-detail/model/lobby-provider";
-import { useAuth } from "@/features/auth/auth-provider";
+import { useAuth } from "@/entities/user/model/auth-provider";
 import type { Chat } from "@/features/lobby/types";
 
 const CHAT_COOLTIME_SECONDS = 300; // 5分
