+++ src/_pages/lobby-detail/ui/lobby-chat-card.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useRef } from "react";
 import { useLobby } from "@/_pages/lobby-detail/model/lobby-provider";
 import { useChat } from "@/_pages/lobby-detail/model/use-chat";
-import { useAuth } from "@/features/auth/auth-provider";
+import { useAuth } from "@/entities/user/model/auth-provider";
 import { Button } from "@/shared/ui/button";
 import { Card, CardContent, CardFooter, CardHeader } from "@/shared/ui/card";
