+++ src/entities/user/model/auth-provider.tsx
